+++ src/main/java/net/specialattack/forge/discotek/client/gui/GuiVerticalSlider.java
@@ -52,7 +52,7 @@ public class GuiVerticalSlider extends GuiSlider {
             }
             minecraft.getTextureManager().bindTexture(Assets.SMALL_GUI);
             GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
-            this.drawTexturedModalRect(this.xPosition + 1, this.yPosition + this.height - 4 - (int) (this.sliderValue * (float) (this.height - 4)), 209, 0, 17, 4);
+            this.drawTexturedModalRect(this.xPosition + 1, this.yPosition + this.height - 4 - (int) (this.sliderValue * (this.height - 4)), 209, 0, 17, 4);
         }
     }
